My new ml model/requirements.txt
flask>=2.2
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=1.7
lightgbm>=4.0
catboost>=1.2
joblib>=1.3
python-dotenv>=1.0
google-generativeai>=0.3.0
